+++ src/main/java/com/sprint/mission/discodeit/data/DataPersistenceManager.java
@@ -1,0 +1,4 @@
+package com.sprint.mission.discodeit.data;
+
+public class DataPersistenceManager {
+}
