+++ src/main/resources/config.properties
@@ -8,4 +8,3 @@ data.path.participations=participations.json
 data.path.channel_messages=channel_messages.json
 data.path.direct_messages=direct_messages.json
 
-# FXML View file paths (from resources root)
