--- src/main/resources/view/LoginView.fxml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Hyperlink?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.PasswordField?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<AnchorPane prefHeight="400.0" prefWidth="600.0" style="-fx-background-color: #36393f;" xmlns="http://javafx.com/javafx/17.0.6" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sprint.mission.discodeit.view.LoginViewController">
-    <VBox alignment="CENTER" layoutX="150.0" layoutY="50.0" spacing="20.0">
-        <Text fill="white" text="Welcome back!">
-            <font>
-                <Font name="System Bold" size="24.0" />
-            </font>
-        </Text>
-        <VBox spacing="8.0">
-            <Label text="USERNAME" textFill="#b9bbbe">
-                <font><Font name="System Bold" size="12.0" /></font>
-            </Label>
-            <TextField fx:id="usernameField" prefWidth="300.0" promptText="Enter your username" style="-fx-background-color: #202225; -fx-text-fill: white; -fx-prompt-text-fill: #72767d;" />
-        </VBox>
-        <VBox spacing="8.0">
-            <Label text="PASSWORD" textFill="#b9bbbe">
-                <font><Font name="System Bold" size="12.0" /></font>
-            </Label>
-            <PasswordField fx:id="passwordField" promptText="Enter your password" style="-fx-background-color: #202225; -fx-text-fill: white; -fx-prompt-text-fill: #72767d;" />
-        </VBox>
-        <Button onAction="#handleLogin" prefWidth="300.0" style="-fx-background-color: #5865f2; -fx-text-fill: white;" text="Login">
-            <font><Font name="System Bold" size="14.0" /></font>
-        </Button>
-        <HBox alignment="CENTER_LEFT">
-            <Text fill="#72767d" text="Need an account? " />
-            <Hyperlink onAction="#handleRegisterLink" text="Register" textFill="#00a8fc" />
-        </HBox>
-        <Label fx:id="errorLabel" textFill="red" />
-    </VBox>
-</AnchorPane>
