--- src/main/resources/view/MainView.fxml
@@ -1,105 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ListView?>
-<?import javafx.scene.control.ScrollPane?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-
-<BorderPane prefHeight="720.0" prefWidth="1280.0" style="-fx-background-color: #36393f;" xmlns="http://javafx.com/javafx/17.0.6" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sprint.mission.discodeit.view.MainViewController">
-    <left>
-        <HBox>
-            <VBox alignment="TOP_CENTER" prefWidth="70.0" spacing="10.0" style="-fx-background-color: #202225;">
-                <padding>
-                    <Insets top="10.0" />
-                </padding>
-                <children>
-                    <Label fx:id="dmButton" alignment="CENTER" onMouseClicked="#handleDmButtonClick" prefHeight="50.0" prefWidth="50.0" style="-fx-background-color: #5865f2; -fx-background-radius: 25;" text="DM" textFill="WHITE">
-                        <font>
-                            <Font name="System Bold" size="14.0" />
-                        </font>
-                    </Label>
-                    <ListView fx:id="serverListView" prefHeight="550.0" style="-fx-background-color: #202225; -fx-padding: 0;" />
-                    <Button onAction="#handleCreateChannel" prefHeight="50.0" prefWidth="50.0" style="-fx-background-color: #313338; -fx-background-radius: 25;" text="+" textFill="LIMEGREEN">
-                        <font>
-                            <Font name="System Bold" size="22.0" />
-                        </font>
-                    </Button>
-                </children>
-            </VBox>
-
-            <VBox prefWidth="240.0" style="-fx-background-color: #2f3136;">
-                <children>
-                    <HBox alignment="CENTER_LEFT" prefHeight="50.0">
-                        <padding>
-                            <Insets left="10.0" right="10.0" />
-                        </padding>
-                        <Label fx:id="subHeaderLabel" text="Direct Messages" textFill="#96989d">
-                            <font>
-                                <Font name="System Bold" size="12.0" />
-                            </font>
-                        </Label>
-                    </HBox>
-                    <ListView fx:id="subListView" style="-fx-background-color: #2f3136;" VBox.vgrow="ALWAYS" />
-
-                    <HBox alignment="CENTER_LEFT" prefHeight="60.0" spacing="10.0" style="-fx-background-color: #292b2f;">
-                        <padding>
-                            <Insets left="10.0" />
-                        </padding>
-                        <VBox>
-                            <Label fx:id="nicknameLabel" text="Nickname" textFill="white">
-                                <font><Font name="System Bold" size="14.0" /></font>
-                            </Label>
-                            <Label fx:id="usernameLabel" text="username" textFill="#96989d" />
-                        </VBox>
-                    </HBox>
-                </children>
-            </VBox>
-        </HBox>
-    </left>
-    <center>
-        <BorderPane>
-            <top>
-                <HBox alignment="CENTER_LEFT" prefHeight="50.0" style="-fx-border-color: #202225; -fx-border-width: 0 0 2 0;">
-                    <padding><Insets left="20.0" /></padding>
-                    <Label fx:id="chatHeaderLabel" text="Welcome to DiscodeIT!" textFill="WHITE">
-                        <font><Font name="System Bold" size="16.0" /></font>
-                    </Label>
-                </HBox>
-            </top>
-            <center>
-                <VBox>
-                    <ScrollPane fx:id="chatScrollPane" fitToWidth="true" style="-fx-background-color: #36393f; -fx-background: #36393f;" VBox.vgrow="ALWAYS">
-                        <content>
-                            <VBox fx:id="chatArea" spacing="15.0">
-                                <padding><Insets bottom="10.0" left="20.0" right="20.0" top="20.0" /></padding>
-                            </VBox>
-                        </content>
-                    </ScrollPane>
-                    <HBox>
-                        <TextField fx:id="messageField" onAction="#handleSendMessage" promptText="Message..." style="-fx-background-color: #40444b; -fx-text-fill: #dcddde; -fx-background-radius: 8;" HBox.hgrow="ALWAYS">
-                            <VBox.margin><Insets bottom="20.0" left="20.0" right="20.0" top="10.0" /></VBox.margin>
-                            <font><Font size="14.0" /></font>
-                        </TextField>
-                    </HBox>
-                </VBox>
-            </center>
-        </BorderPane>
-    </center>
-    <right>
-        <VBox fx:id="memberListPanel" prefWidth="240.0" style="-fx-background-color: #2f3136;">
-            <children>
-                <Label prefHeight="50.0" text="  MEMBERS" textFill="#96989d">
-                    <padding><Insets left="10.0" /></padding>
-                    <font><Font name="System Bold" size="12.0" /></font>
-                </Label>
-                <ListView fx:id="userListView" style="-fx-background-color: #2f3136;" VBox.vgrow="ALWAYS" />
-            </children>
-        </VBox>
-    </right>
-</BorderPane>
