--- src/main/resources/view/RegisterView.fxml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Hyperlink?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.PasswordField?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<AnchorPane prefHeight="500.0" prefWidth="600.0" style="-fx-background-color: #36393f;" xmlns="http://javafx.com/javafx/17.0.6" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sprint.mission.discodeit.view.RegisterViewController">
-    <VBox alignment="CENTER" layoutX="150.0" layoutY="30.0" spacing="15.0">
-        <Text fill="white" text="Create an account">
-            <font><Font name="System Bold" size="24.0" /></font>
-        </Text>
-        <TextField fx:id="emailField" prefWidth="300.0" promptText="Email" style="-fx-background-color: #202225; -fx-text-fill: white;" />
-        <TextField fx:id="usernameField" promptText="Username" style="-fx-background-color: #202225; -fx-text-fill: white;" />
-        <PasswordField fx:id="passwordField" promptText="Password" style="-fx-background-color: #202225; -fx-text-fill: white;" />
-        <TextField fx:id="nicknameField" promptText="Nickname" style="-fx-background-color: #202225; -fx-text-fill: white;" />
-        <TextField fx:id="phoneNumField" promptText="Phone Number (Optional)" style="-fx-background-color: #202225; -fx-text-fill: white;" />
-        <Button onAction="#handleRegister" prefWidth="300.0" style="-fx-background-color: #5865f2; -fx-text-fill: white;" text="Continue">
-            <font><Font name="System Bold" size="14.0" /></font>
-        </Button>
-        <Hyperlink onAction="#handleBackToLogin" text="Already have an account?" textFill="#00a8fc" />
-        <Label fx:id="errorLabel" textFill="red" />
-    </VBox>
-</AnchorPane>
